move helpers.go to utils.go
